each metric gets a chart, searcher metric first
