+++ webui/react/src/pages/TrialDetails/TrialCharts.module.scss
@@ -1,7 +1,6 @@
 .base {
   display: flex;
   flex-direction: column;
-  height: 400px;
   justify-content: center;
 }
 .spinner {
@@ -9,4 +8,8 @@
   display: flex;
   justify-content: center;
   width: 100%;
+}
+.base h5 {
+  align-items: center;
+  text-align: center;
 }
